requirements.txt
beautifulsoup4~=4.12.3
Requests~=2.32.3
speedtest_cli~=2.1.3
tqdm~=4.66.5
